Move ProtoClient to RPC crate, behind feature flag disabled in collab
