+++ crates/rpc/Cargo.toml
@@ -14,6 +14,8 @@ path = "src/rpc.rs"
 doctest = false
 
 [features]
+default = ["client"]
+client = ["gpui"]
 test-support = ["collections/test-support", "gpui/test-support", "proto/test-support"]
 
 [dependencies]
@@ -38,7 +40,6 @@ zstd = "0.11"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 zstd = { version = "0.11", features = [ "pkg-config" ] }
-
 
 [dev-dependencies]
 collections = { workspace = true, features = ["test-support"] }
