+++ crates/rpc/src/rpc.rs
@@ -14,4 +14,9 @@ pub use peer::*;
 pub use proto::{error::*, Receipt, TypedEnvelope};
 mod macros;
 
+#[cfg(feature = "client")]
+mod proto_client;
+#[cfg(feature = "client")]
+pub use proto_client::*;
+
 pub const PROTOCOL_VERSION: u32 = 68;
